// src/types/api.ts
export interface PokemonListItem {
    name: string ;
    url: string; 
}

export interface PokemonListResponse {
    count: number;
    next: string| null ; 
    previous: string| null; 
    results: PokemonListItem[]; 
}
export interface PokemonDetail {
    id: number;
    name: string;
    height: number;
    weight: number;
    sprites: {
      front_default: string;
    };
    types: {
      type: {
        name: string;
      };
    }[];
  }